format all code.

+++ zuihou-msgs/zuihou-msgs-biz/src/main/java/com/github/zuihou/msgs/service/impl/MsgsCenterInfoServiceImpl.java
@@ -37,18 +37,21 @@ import static com.github.zuihou.utils.StrHelper.getOrDef;
  */
 @Slf4j
 @Service
-public class MsgsCenterInfoServiceImpl extends SuperServiceImpl<MsgsCenterInfoMapper, MsgsCenterInfo> implements MsgsCenterInfoService {
+public class MsgsCenterInfoServiceImpl extends SuperServiceImpl<MsgsCenterInfoMapper, MsgsCenterInfo> implements MsgsCenterInfoService
+{
     @Autowired
     private MsgsCenterInfoReceiveService msgsCenterInfoReceiveService;
 
     @Override
-    public IPage<MsgsCenterInfoPageResultDTO> page(IPage<MsgsCenterInfoPageResultDTO> page, MsgsCenterInfoQueryDTO data) {
+    public IPage<MsgsCenterInfoPageResultDTO> page(IPage<MsgsCenterInfoPageResultDTO> page, MsgsCenterInfoQueryDTO data)
+    {
         return baseMapper.page(page, data);
     }
 
     @Override
     @Transactional(rollbackFor = Exception.class)
-    public MsgsCenterInfo saveMsgs(MsgsCenterInfoSaveDTO data) {
+    public MsgsCenterInfo saveMsgs(MsgsCenterInfoSaveDTO data)
+    {
         MsgsCenterInfo info = BeanPlusUtil.toBean(data.getMsgsCenterInfoDTO(), MsgsCenterInfo.class);
         info.setTitle(getOrDef(info.getTitle(), info.getContent()));
         info.setAuthor(getOrDef(info.getAuthor(), BaseContextHandler.getName()));
@@ -56,12 +59,11 @@ public class MsgsCenterInfoServiceImpl extends SuperServiceImpl<MsgsCenterInfoMa
 
         //公式公告，不会指定接收人
         Set<Long> userIdList = data.getUserIdList();
-        if (CollectionUtil.isNotEmpty(userIdList)) {
-            List<MsgsCenterInfoReceive> receiveList = userIdList.stream().map((userId) -> MsgsCenterInfoReceive.builder()
-                    .isRead(false)
-                    .userId(userId)
-                    .msgsCenterId(info.getId())
-                    .build()).collect(Collectors.toList());
+        if (CollectionUtil.isNotEmpty(userIdList))
+        {
+            List<MsgsCenterInfoReceive> receiveList = userIdList.stream()
+                                                                .map((userId) -> MsgsCenterInfoReceive.builder().isRead(false).userId(userId).msgsCenterId(info.getId()).build())
+                                                                .collect(Collectors.toList());
             msgsCenterInfoReceiveService.saveBatch(receiveList);
         }
         return info;
@@ -69,16 +71,16 @@ public class MsgsCenterInfoServiceImpl extends SuperServiceImpl<MsgsCenterInfoMa
 
     @Override
     @Transactional(rollbackFor = Exception.class)
-    public boolean delete(List<Long> ids, Long userId) {
-        msgsCenterInfoReceiveService.remove(Wraps.<MsgsCenterInfoReceive>lbQ()
-                .eq(MsgsCenterInfoReceive::getUserId, userId)
-                .in(MsgsCenterInfoReceive::getMsgsCenterId, ids));
+    public boolean delete(List<Long> ids, Long userId)
+    {
+        msgsCenterInfoReceiveService.remove(Wraps.<MsgsCenterInfoReceive>lbQ().eq(MsgsCenterInfoReceive::getUserId, userId).in(MsgsCenterInfoReceive::getMsgsCenterId, ids));
 
-        for (Long msgCenterId : ids) {
-            int count = msgsCenterInfoReceiveService.count(Wraps.<MsgsCenterInfoReceive>lbQ()
-                    .eq(MsgsCenterInfoReceive::getMsgsCenterId, msgCenterId));
+        for (Long msgCenterId : ids)
+        {
+            int count = msgsCenterInfoReceiveService.count(Wraps.<MsgsCenterInfoReceive>lbQ().eq(MsgsCenterInfoReceive::getMsgsCenterId, msgCenterId));
             log.info("count={}", count);
-            if (count <= 0) {
+            if (count <= 0)
+            {
                 super.remove(Wraps.<MsgsCenterInfo>lbQ().eq(MsgsCenterInfo::getId, msgCenterId));
             }
         }
@@ -97,22 +99,23 @@ public class MsgsCenterInfoServiceImpl extends SuperServiceImpl<MsgsCenterInfoMa
      */
     @Override
     @Transactional(rollbackFor = Exception.class)
-    public boolean mark(List<Long> msgCenterIds, Long userId) {
-        if (CollectionUtil.isEmpty(msgCenterIds) || userId == null) {
+    public boolean mark(List<Long> msgCenterIds, Long userId)
+    {
+        if (CollectionUtil.isEmpty(msgCenterIds) || userId == null)
+        {
             return true;
         }
         List<MsgsCenterInfo> list = super.listByIds(msgCenterIds);
 
         //其他类型的修改状态
-        if (!list.isEmpty()) {
+        if (!list.isEmpty())
+        {
             List<Long> idList = list.stream().mapToLong(MsgsCenterInfo::getId).boxed().collect(Collectors.toList());
-            msgsCenterInfoReceiveService.update(Wraps.<MsgsCenterInfoReceive>lbU()
-                    .eq(MsgsCenterInfoReceive::getUserId, userId)
-                    .in(MsgsCenterInfoReceive::getMsgsCenterId, idList)
-                    .set(MsgsCenterInfoReceive::getIsRead, true)
-                    .set(MsgsCenterInfoReceive::getUpdateTime, LocalDateTime.now())
-                    .set(MsgsCenterInfoReceive::getUpdateUser, userId)
-            );
+            msgsCenterInfoReceiveService.update(Wraps.<MsgsCenterInfoReceive>lbU().eq(MsgsCenterInfoReceive::getUserId, userId)
+                                                                                  .in(MsgsCenterInfoReceive::getMsgsCenterId, idList)
+                                                                                  .set(MsgsCenterInfoReceive::getIsRead, true)
+                                                                                  .set(MsgsCenterInfoReceive::getUpdateTime, LocalDateTime.now())
+                                                                                  .set(MsgsCenterInfoReceive::getUpdateUser, userId));
         }
         return true;
     }
